wip: code cleanup
